refactor reference, remove run reference
