+++ BALSAMIC/commands/init/reference.py
@@ -7,9 +7,7 @@ import graphviz
 import snakemake
 from pathlib import Path
 
-from BALSAMIC.utils.cli import write_json, merge_json
-from BALSAMIC.utils.cli import get_snakefile
-from BALSAMIC.utils.cli import CaptureStdout
+from BALSAMIC.utils.cli import write_json, merge_json, CaptureStdout, get_snakefile, SnakeMake
 from BALSAMIC import __version__ as bv
 
 LOG = logging.getLogger(__name__)
@@ -46,9 +44,27 @@ LOG = logging.getLogger(__name__)
               type=click.Choice(["hg19", "hg38"]),
               help=("Genome version to prepare reference. Path to genome"
                     "will be <outdir>/genome_version"))
+@click.option(
+    '-r',
+    '--run-analysis',
+    show_default=True,
+    default=False,
+    is_flag=True,
+    help='By default balsamic run_analysis will run in dry run mode. \
+              Raise thise flag to make the actual analysis')
+@click.option(
+    '-f',
+    '--force-all',
+    show_default=True,
+    default=False,
+    is_flag=True,
+    help='Force run all analysis. This is same as snakemake --forceall')
+@click.option('--snakemake-opt',
+              multiple=True,
+              help='Pass these options directly to snakemake')
 @click.pass_context
 def reference(context, outdir, cosmic_key, snakefile, dagfile, singularity,
-              genome_version):
+              genome_version, run_analysis, force_all, snakemake_opt):
     """ Configure workflow for reference generation """
 
     LOG.info(f"BALSAMIC started with log level {context.obj['loglevel']}.")
@@ -109,3 +125,29 @@ def reference(context, outdir, cosmic_key, snakefile, dagfile, singularity,
     except Exception:
         LOG.error(f'Reference workflow graph generation failed')
         raise click.Abort()
+
+    LOG.info(f"BALSAMIC started with log level {context.obj['loglevel']}.")
+    LOG.info("Reference generation workflow started")
+
+    # Singularity bind path
+    bind_path = list()
+    bind_path.append(config['output'])
+    bind_path.append(config['conda_env_yaml'])
+    bind_path.append(config['rule_directory'])
+
+    # Construct snakemake command to run workflow
+    balsamic_run = SnakeMake()
+    balsamic_run.working_dir = config['output']
+    balsamic_run.snakefile = snakefile
+    balsamic_run.configfile = config_json
+    balsamic_run.run_mode = "local"
+    balsamic_run.forceall = force_all
+    balsamic_run.run_analysis = run_analysis
+    balsamic_run.sm_opt = snakemake_opt
+
+    # Always use singularity
+    balsamic_run.use_singularity = True
+    balsamic_run.singularity_bind = bind_path
+    balsamic_run.sm_opt = snakemake_opt
+
+    subprocess.run(balsamic_run.build_cmd(), shell=True)
